Android-21-CheckNet
Check the current network status. When there is a network status, show-"A network is available". If there is no network, a message window will appear, indicating that there is no network. Do you want to go to the settings? Select the "Setup Network" item, jump to the system's wifi setting page, select "No setting", and close the window.